requirements.txt
    transformers
    sentence-transformers
    faiss-cpu
    langchain
    fastapi
    uvicorn